plasm v0.1.0 Plasm.Database.MySql

Functions

connect(conn_string, opts \\ [])

Connect to the database

convert_to_utf8(raw)

create_connection_string(server, port \\ 3306, user_name, password, version, database_name)

Create the connection string specific for this version of MySql. This will require the MySql ODBC driver to be installed on the computer in order to run

get_table_fields(conn, table)

Get all of the columns and their types from the database

get_tables(conn, query)

Get all of the tables from the given database

tables(conn, db_name)

Get the tables from the specific database. We get a list of tuple/1 so we need to get the first element of that tuple then convert it to UTF-8 from UTF-16 in order

to be able to actual retreive the file name. Elixir uses UTF8 strings so if we don't do the conversion that means we will have errors.'