plasm v0.1.0 Plasm.Export

Functions

output_table(table, fields, project_name)

Format the text of a specific table with the fields that are passed in. This is strictly formatting and will not verify the fields with the database

type_output(field)

Generate the schema field based on the database type