plsm v1.1.2 Plsm.IO.Export

Functions

prepare(table, project_name)

prepare(Plsm.Database.Table, String.t) :: String.t

Format the text of a specific table with the fields that are passed in. This is strictly formatting and will not verify the fields with the database

type_output(field)

Generate the schema field based on the database type

write(schema, name, path \\ "")

write(String.t, String.t, String.t) :: Any

Write the given schema to file.