plsm v2.3.0 Plsm.Database protocol

Types

t()

t() :: term()

Functions

connect(db)

Connect to the given database

create(db, configs)

Create a database struct that implements the Plsm.Database protocol.

get_columns(db, table)

Get the columns for the table that is passed in

get_tables(db)

Get all of the tables that are in the database that was selected